Extract participations section (public)
